tracking instance no

+++ src/main/java/driver/Pruning.java
@@ -25,24 +25,24 @@ public class Pruning {
      */
     public static boolean getMDL(final Gain gain) {
 
-        final float gainValue = gain.getGain();
-        final float entropyA = gain.getEntropyA();
-        final float entropyB = gain.getEntropyB();
-        final float entropyAB = gain.getEntropy();
+        final double gainValue = gain.getGain();
+        final double entropyA = gain.getEntropyA();
+        final double entropyB = gain.getEntropyB();
+        final double entropyAB = gain.getEntropy();
 
-        final float N = gain.indexListA.size() + gain.indexListB.size();
-        final float A = gain.occurrenceA.size();
-        final float B = gain.occurrenceB.size();
+        final double N = gain.indexListA.size() + gain.indexListB.size();
+        final double A = gain.occurrenceA.size();
+        final double B = gain.occurrenceB.size();
         // |AuB| - number of possible class labels in entire set
         final TreeSet<String> classesNames = new TreeSet<String>();
         final Set a = gain.occurrenceA.keySet();
         final Set b = gain.occurrenceB.keySet();
         classesNames.addAll(a);
         classesNames.addAll(b);
-        final float AuB = classesNames.size();
+        final double AuB = classesNames.size();
 
-        final float leftSideOfFormula = ((1 / N) * (float) (Math.log(N - 1) / Math.log(2))) +
-                (1 / N) * ((float) (Math.log(Math.pow(3, AuB) - 2) / Math.log(2)) - (AuB * entropyAB) - (A * entropyA) - (B * entropyB));
+        final double leftSideOfFormula = ((1 / N) * (Math.log(N - 1) / Math.log(2))) +
+                (1 / N) * ((Math.log(Math.pow(3, AuB) - 2) / Math.log(2)) - (AuB * entropyAB) - (A * entropyA) - (B * entropyB));
         final boolean isSplit = (gainValue >= leftSideOfFormula);
         //System.out.println("MDL: " + leftSideOfFormula + "\nsplit:" + isSplit + "\nThreshold " + gain.getThreshold() + "\n");
         return gainValue >= leftSideOfFormula;
